refactoring the code according to the remarks

+++ src/main/java/com/shop/model/DraftBeerData.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class DraftBeerData implements BeerInfo {
-    private Integer availableLiters;
+    private Double availableLiters;
 }
